y_previous_attempts/y_old_requirements.OLD.txt
# Core agent framework (pre-release)
# Note: agent-framework-core must stay at b260107 because agentserver b14
# constrains it below rc1 (AgentProtocol/AIFunction removed in rc1).
agent-framework==1.0.0rc1
agent-framework-core==1.0.0b260107
agent-framework-orchestrations==1.0.0b260219

# Azure agentserver adapter — wraps workflow as HTTP server (port 8088)
azure-ai-agentserver-core==1.0.0b14
azure-ai-agentserver-agentframework==1.0.0b14

# OpenTelemetry semconv — 0.4.13 exposes SpanAttributes.LLM_SYSTEM
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-exporter-otlp-proto-grpc==1.39.0

# Azure auth + env
azure-identity==1.26.0b2
python-dotenv==1.2.1
